+++ R/unmixing_quality_addin.R
@@ -56,7 +56,7 @@ addin_40_unmixing_quality = function() {
     # Handle the browse button by selecting a folder
     shiny::observeEvent(input$browse, {
       shiny::req(input$browse)
-      export_dir(utils::choose.dir(
+      export_dir(phenoptrReports::choose_directory(
         caption='Select an export folder'
       ))